return success if the existing file has correct contents
If we fail to rename over the old file, compare the files.

If the new file we are trying to rename over the old one
has identical contents, we can still call it a success.

This can happen in various scenarios, possibly with
prepopulated cache directories or different users sharing
the same directory.